Try to address flaky database transaction tests. (#235)
The test "Transaction without local events (2)" is inherently racey (we're trying to do 4 sets on one connection before the other connection retries the transaction 25 times). And when it fails, it could cause the rest of the tests to fail by not calling restoreHash(). So I've disabled the test and made the restoreHash handling a bit more robust so similar cascading failures can't happen in the future.