Upgrade ts-node and disable its cache. (#446)
I upgraded ts-node while investigating a cache-related issue and that solved my issue, but it also led me to realize that the cache is inaccurate and can hide type failures.

By default ts-node caches the JS output for each TS file it encounters. But if A.ts depends on types in B.ts and B.ts changes, ts-node will still use the cached compilation of B.ts even if it no longer should compile.

Disabling the cache adds ~6s to incremental test time but ensures we don't miss type violations. It should have no effect on travis which always starts with a clean cache.