Add an option to settings to control how timestamp fields are returned.
The behavior to return timestamp fields as dates is now legacy. It's on
by default to allow for a transitional period. The new behavior to
return Timestamp objects instead is opt-in, but will become default in
the future. The advantage Timestamps have is higher precision than
native browser Date. If the new behavior is not enabled, a warning is
logged on startup urging users to upgrade.