Remove "google-cloud-resource-prefix" header in favor of "database" url parameter in WebChannel requests.
I verified that the client still works against prod and that the ?database=... paremeter was included in every WebChannel network request.

NOTE that prior to this change we were including the "google-cloud-resource-prefix" header on non-streaming RPC requests, but with this change we will not include the header or use the "database" url parameter.