Adds tests in demo app for service worker compatibility with Firebase Auth.
This modifies the service worker to append the ID token in every fetch
request via a custom HTTP header.

Using Functions with Hosting, an HTTP endpoint was setup which will
parse the ID token from the header and verify it using Admin SDK and
then return a status response.

A button is added to the demo app to test this. When a user is signed
in, the user's uid is returned, otherwise an error is shown.