Small cleanup to FieldValue sentinel code to make it easier to add more sentinels in the future.
This technically includes a small breaking change. Before we would allow you to do:

doc.set({a: [0, FieldValue.delete(), 2] }, {merge: true})

Which was equivalent to:

doc.set({a: [0, null, 2] }, {merge: true})

But with my change it will now result in an error:

FieldValue.delete() is not currently supported inside arrays.