Add karma config for running all tests with SauceLabs.
IE11 is not supported due to the compatibility issue. It will be
enabled once the issue is resolved.

Currently these tests are excluded for saucelabs testing:

  'packages/database/*',
  'packages/firestore/*',
  'packages/messaging/*',
  'integration/firestore/*',
  'integration/messaging/*'.

Database and firestore tests appear to be flaky on saucelabs and
tend to take much longer time than running in local browsers. Message
tests use service worker functionality which is not supported by Safari
and Edge currently. They will be included later on when these problems
are resolved.