Expand browsers support (#674)
* Add karma config for running all tests with SauceLabs.

IE11 is not supported due to the compatibility issue. It will be
enabled once the issue is resolved.

Currently these tests are excluded for saucelabs testing:

  'packages/database/*',
  'packages/firestore/*',
  'packages/messaging/*',
  'integration/firestore/*',
  'integration/messaging/*'.

Database and firestore tests appear to be flaky on saucelabs and
tend to take much longer time than running in local browsers. Message
tests use service worker functionality which is not supported by Safari
and Edge currently. They will be included later on when these problems
are resolved.

* Group Travis jobs for different test commands into one test stage.

Deployment starts only after all jobs in test stage have succeeded.

* Add karma Firefox launcher to each applicable package.

In case of a pull request build on Travis where saucelabs testing
is not available, tests will run locally in Chrome and Firefox on
Travis VMs instead.

https://docs.travis-ci.com/user/pull-requests/#Pull-Requests-and-Security-Restrictions

* fix nit