Remove startsWith because it's not supported in IE11
Make IE great again.

I tried adding the tslint rule `"no-unsupported-browser-code": [true, ["IE 11"]],` but as it turns out the package `database` doesn't lint the code.
Linting is only enabled in `firestore` and `messaging`

I went a bit deeper trying to copy the config from `firestore` but as it turns out, there are loads of lint errors and I don't feel like touching more code than necessary.

For the moment simply remote the unsupported method in IE11, if anyone cares please add linting and configure it so that these errors won't show up again in the future