Fixes service worker indexedDB on some slow browsers, which was detected in Edge.
Unfortunately, some browsers (detected in Edge) are too slow to detect the first ack event and timeout quickly.
Instead what we can do is assume the following:
1. one service worker activated at one time.
2. The service worker once subscribed to the keyChanged event will not unsubscribe.
So on initialization, the client will ping the service worker when it's ready to get all subscribed events with a long timeout.
If the service worker responds with the keyChanged event, we can deduce that the sw can handle this event.
The next time a change event occurs, we can safely use a long ACK timeout.