Returns the credential when available in AuthErrorWithCredential
This returns credentials in errors even when no email or phoneNumber is returned.
This is needed to help recover from credential auth/credential-already-in-use in multiple accounts per email settings, eg: user calling linkWithPopup/Redirect when the OAuth credential is already in use.
In this case, the top level email is null and the developer can't do account lookup but at least they can still recover. The developer can signInWithCredential after and handle merges manually.