Fix flaky database transaction test.
The test was calling done() while there were still outstanding transactions,
and a subsequent test was calling goOffline(), which then caused the
outstanding transactions to be aborted, leading to the assertion in the
completion callback to fail. This probably only reproduced with the emulator
because it has much less latency than prod.