Overhaul Documents
Remove different internal types for MaybeDocument, Document, NoDocument,
and UnknownDocument. These are all now just Documents.

Also make Documents non-nullable nearly everywhere.

The change introduces a DocumentType enum and maps the old classes like
so:

  null =>             DocumentType.UNKNOWN
  UnknownDocument =>  DocumentType.CONTENTS_UNKNOWN
  NoDocument =>       DocumentType.MISSING
  Document =>         DocumentType.EXISTING

Part of why this change works is that most of the system only cares
whether or not a document exists. We have several refinements of
non-existing documents for the purposes of caching but everywhere else
we can ignore it and just check Document.exists. The instanceof checks
were usually overconstraining things.

The motivation for this change is in two parts:

  * It's always bugged me that we had so many instanceof checks around
  * In C++ we don't have instanceof anyway, so we have to do something
    else.

Also, this definitely addresses b/32275378.