Make spec tests easier to debug
With this simple change, tests not marked as not "exclusive" will not be translated into JSON. Hence, we can set breakpoints in the spec builder functions. It's really a win win.

I also expect this to speed up our spec tests so much that this changes the way we look at unit testing. I am talking microseconds.