Avoid using $httpHeaders in environments that don't send an Origin header.
There is a backend bug (b/145624756) causing $httpHeaders to be ignored if an Origin header is not sent.  Via https://github.com/firebase/firebase-js-sdk/issues/1491 we've found that several environments can run into this issue so until the backend issue is fixed we are excluding them from using $httpHeaders. This may incur an extra CORS preflight round-trip in some cases, but this is much better than having our Authorization header get ignored.