Debug and release asserts
This PR aims to differentiate between asserts that we should handle in production builds (e.g. we received invalid data) and asserts we handle only during testing (our local state is corrupt). Some of the distinctions are really obvious, and other are anything but. In a lot of cases, we actually have multiple asserts for a single condition. I opted to only  make the initial assertion a hardAssert (e.g. when we reveive mutation results, we assert multiple times that the number of results matches the mutation count).

softAssert is a temporary name that  I will rename back to assert once this is done. It is meant to make the decisions here more obvious