Ignore IndexedDb failures during LRU
Rationale is that they the LRU task will retry its run when it is invoked next.

There are no tests since the GC scheduling code is not tested right now. Let me know if you would like me to add tests.

Note that there is a behavior change: If the primary lease is lost, the GC run is rescheduled. It will however get stoped in
https://github.com/firebase/firebase-js-sdk/blob/5a60243bf264967819fc5c3897892084c5eb4d43/packages/firestore/src/core/component_provider.ts#L203