Specify readonly as the default indexeddb transaction mode
While the mode parameter is optional, the wrapper method will pass `undefined` into the native method if it is not passed into the wrapper. IE11 interprets this as a mode instead of ignoring it and throws an InvalidAccessError.

Specifying `readonly` as the default (which is the default if the parameter is omitted) prevents IE11 from crashing during initialization.