Fix types for setUserId
The types for setUserId in the service module did not allow for passing
null, which is the recommended way for clearing out the user id:
https://firebase.google.com/docs/reference/android/com/google/firebase/analytics/FirebaseAnalytics#public-void-setuserid-string-id

I updated the types and related documentation to allow null to be passed
and added a test. The underlying setUserId function in the functions
module already allows for null, so this shouldn't result in any
behavioral changes.