Enable encodeInitMessageHeaders (#6107)
* Enable encodeInitMessageHeaders

This transitions the Firestore client from encoding HTTP Headers via the Query Param to the request's POST payload.

* Create eleven-moons-shave.md

Set the change type to 'minor'. There is no change to the public API and no feature added, but the change does require an update to the emulator. We feel that a patch may go unnoticed.

* Update firestore emulator used by tests to v1.14.4