Expose TOKEN_EXPIRED error upon mfa unenroll.
This can be thrown if the MFA option that was most recently enrolled into, was unenrolled.
The user will be logged out to prove the posession of the other second factor.
This error can be handled by reauthenticating the user.

This change also updates the demo app to store the lastUser in case mfa unenroll logs out the user.
From here, the lastUser can be reauthenticated.