Wait for upto 8s for IDP sign in to finish, after popup is closed.
In some cases (Fiefox or mobile or if opener is an iframe), the popup is closed by the oauth helper code
right after sign in has completed at the IDP. The IDP response still needs to be processed by the SDK + signInWithIdp API
call needs to be invoked. This can take upto 7s, if there is a blocking function configured. Increase the poller timeout
to 8s, so it does not reject the sign in with popup-closed-by-user error.

increase timeout for the aborted sign in test.