Fix to test-all .github workflow yaml file. (#7640)
CI Changes only: Fixed a misaligned indentation on line #107 of test-all.yaml that cause a workflow execution failure in GitHub Action CI.

Includes an update to all of the test-all.yml chrome installation steps to use the npx install method.

There's a Failure in (Firestore) Node.js and Browser (Chrome) Tests, but this existed even before the npx update (when installing Chrome via apt-get). See this previous run. Since the Firestore issue still remains, the has been disabled (will skip) for now.

Also fixed the a problem setting the max heap size for node. Previous attempts echoed the value to the GITHUB_ENV variable on the command line, but this was producing errors. Changed it to use the worfklow env: directive and I have confirmed that the enviornment variable exists in the shell runtime (example echo output can be found in run unit tests step of this worklfow run).