Fix E2E runtime - pin Firebase Tools, node 18 runtime (#7865)
Updates to our E2E tests in CI (`E2E Smoke Tests` and `Deploy Project Config`) to use Node v18. Additionally, pin the firebase-tools version to 12.9.1. While we plan to upgrade the firebase-tools version in the future, we're chosing to pin to a version for the long term to prevent suddent CI test failures.