[E2E CI] Update firebase-tools to v13.0.2 (#7883)
Previously we ran into an issue where we couldn't use the latest firebase-tools CLI version (v13+) due to an error attempting to deploy the E2E cloud function.

The error was due to a new requirement of the CLI that `firebase.json` requires a `functions` configuration. This change adds that configuration and pins the firebase CLI version to 13.0.2.