Enable previously failing `(Firestore) Node.js and Browser (Chrome) Tests` (#8067)
A Firestore CI test configuration was previously disabled in our GitHub action Workflows due to a 100% failure rate. While revisiting this issue months later, it seems that the test now passes. This might be related to our Node v20 CI upgrade.

As a bonus, fixed a comment typo in the check changeset script.