Use GitHub ref to get PR head
The pull request head ref name is local to its originating repository.
For pull requests from forks, that ref is unlikely to exist.

GitHub has special magic refs (refs/pull/<number>/<head|merge>) that can
be used to reference the pull request from the destination repository's
perspective.

Note that the merge magic ref is not instance and is slightly quirky.
Here, we're using the head ref which is less quirky.