Run Storage Node tests in CI
The script that runs the tests in CI accepts a `-s` flag we use to pass
the test command we want to run in CI. We were passing two commands, and
only the first one (`test:browser`) was being run, skipping the Node
tests. Now we only pass one command `test`, which runs both.