Consolidate CI test output into a single string
Having CI test process stdout and stderr in a single string makes it
easier to read when looking through failures, since you can see the test
output alongside the error messages.

Without this, any stderr output written during a test will be captured
seperately from the test output, so when we then log it after a test
failure, we can't tell which test logged which errors.