Remove `files` from all karma module exports.
This object is not used by Karma, so I believe there is no need to
export it. Karma instead uses the 'files' property that is assigned on the
config object.

Running tests locally, it seems that the same set of tests are ran
after this change.

One side-effect of this change could be that files importing the another
`karma.conf.js` file may use these exported files, but I can not find
cases of this.