test(vertexai): add `backendName` param to mock response getters (#8906)
`backendName` can be either `googleAI` or `vertexAI`. This can be passed
to `getMockResponse` or `getMockResponseStreaming` to specify whether
the mock response lookup should be done in the set of vertexAI or
googleAI mock files.

Modified the `convert-mocks.ts` script to read mock responses from the
'developerapi' directory, and add an export to the generated file for
the new lookup object with those mock responses.