[FEAT] Restrict Firestore Doc/QuerySnapshot toJSON on client. (#9048)
To reduce bundle size on clients, the `toJSON` methods of `DocumentSnapshot` and `QuerySnapshot` will not create bundles.

### Testing

Updated the integration and unit tests to test `fromJSON` functionality in node environments only.

I did attempt to create pre-packaged bundles so that we can test in browser still, but unforunately the bundle includes the project name, and the project that we test against varies depending on the profile of tests that were running (prod, local, etc). If these don't match then an error is thrown.

### API Changes

N/A