fix(ai): Only include unexpected prediction element in error message.
If the `responseJson` from the Predict request has an image, the entire
image base64 will be included in the error message. This can make the error
message >400kB (!!), which could cause performance issues in apps.

I ran into this when testing https://github.com/firebase/firebase-js-sdk/pull/9216,
since the `safetyAttributes` would be included in a response and trigger
this error, but the `predictions` array in `responseJson` would have images.
I noticed that when the sample app rendered the `error.message`, it
began running into serious performance issues.