fix(ai): remove error logging in `sendMessageStream()`
Removes error logging in `sendMessageStream()`. These errors are
catchable by the user, so there is no need to log them.
This removes the anti-pattern of throwing a `SILENT_ERROR` to skip
promise chains.

This fix is important for when we add [AbortSignal](
https://github.com/firebase/firebase-js-sdk/pull/8890). Since aborting a
request raises an `AbortError`, every time a request is aborted, there
will be an error logged to the console, which is undesirable for the
user.