test(ai): Remove flaky token count integration expect directives (#9401)
We had some "ballpark" checks for the number of tokens returned in responses. 

However, the number of tokens returned had a high variance based on general AI chaos and the model that was used. This PR replaces somewhat exacting token count checks with simple non-zero checks in order to reduce flake.

Additionally, the change removes `toolUsePromptTokenCount` checks for the `gemini-2.0-flash` model, which doesn't support this field.